\name{OptionalNumeric-class}
\docType{class}
\alias{OptionalNumeric-class}

\title{Optional numeric}
\description{Optional object of class \code{"numeric"}.}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
%\section{Methods}{}
%\references{}
\author{Matthias Kohl \email{Matthias.Kohl@stamats.de}}
%\note{}
\seealso{\code{\link[methods:BasicClasses]{numeric-class}}}
\keyword{classes}
